examples/requirements.txt
# Requirements for NOOB Examples
# Install with: pip install -r requirements.txt

# Core dependencies
numpy>=1.20.0
Pillow>=9.0.0

# Optional: For distributed execution
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0

# Optional: For blockchain integration
web3>=6.0.0
eth-account>=0.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
